Se ha agregado estilos a el navbar, los enlaces de las redes sociales y ajustado el sistema de carga de archivos, pero todavia falta depurar errores en la carga de las imagenes

+++ src/redux/actions/projectActions.js
@@ -1,6 +1,6 @@
 import { createAction } from 'redux-actions';
 import config from '../../config';
-import { projectHttp } from '../../service/fetch';
+import { projectHttp, uploadFile } from '../../service/fetch';
 
 const url = config.url+"/proyect/";
 
@@ -19,8 +19,11 @@ export const getAll = () => async (dispatch) => {
 export const deleteProjectSuccess = createAction('DELETE_AN_PROJECT');
 export const deleteProject = (id) => async (dispatch) => {
     try{
-        projectHttp('DELETE', `remove/${id}`, null);
-        dispatch(deleteProjectSuccess(id));
+        const confirm = window.confirm('Esta seguro que desea eliminar este proyecto?');
+        if(confirm){
+            projectHttp('DELETE', `remove/${id}`, null);
+            dispatch(deleteProjectSuccess(id));
+        }
     }
     catch(error){
         console.log(error);
@@ -30,7 +33,18 @@ export const deleteProject = (id) => async (dispatch) => {
 export const addNewProjectSuccess = createAction('CREATE_NEW_PROJECT');
 export const addNewProject = (project) => async (dispatch) => {
     try{
-        const response = await projectHttp('POST', `add`, project);
+        console.log(project);
+        let response = null;
+        if(project.image !== null){
+            console.log(project.image);
+            const img = project.image;
+            project.image = null;
+            response = await projectHttp('POST', `add`, project);
+            response = await uploadFile(response._id, img);
+        }
+        else{
+            response = await projectHttp('POST', `add`, project);
+        }
         dispatch(addNewProjectSuccess(response));
     }
     catch(error){
@@ -39,9 +53,20 @@ export const addNewProject = (project) => async (dispatch) => {
 }
 
 export const updateProjectSuccess = createAction('UPDATE_PROJECT');
-export const updateProject = (project) => async (dispatch) => {
+export const updateProject = (project, changeImg) => async (dispatch) => {
     try{
-        await projectHttp('PUT', `update/${project._id}`, project);
+        console.log(changeImg);
+        if(changeImg === true){
+            console.log("true");
+            const img = project.image;
+            project.image = null;
+            await projectHttp('PUT', `update/${project._id}`, project);
+            const response = await uploadFile(project._id, img);
+            project.image = response.image;
+        }else{
+            console.log("false");
+            await projectHttp('PUT', `update/${project._id}`, project);
+        }
         dispatch(updateProjectSuccess(project));
     }
     catch(error){
